pyproject.toml
[project]
name = "klawtrade"
version = "0.1.0"
description = "Autonomous algorithmic trading system with rule-based strategies, risk management, and real-time dashboard"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
keywords = ["trading", "algorithmic-trading", "paper-trading", "alpaca", "finance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "aiosqlite>=0.22.1",
    "alpaca-py>=0.38.0",
    "fastapi>=0.135.2",
    "jinja2>=3.1.6",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pandas-ta>=0.4.71b0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
    "websockets>=16.0",
]

[project.scripts]
klawtrade = "src.cli:cli"

[project.urls]
Homepage = "https://github.com/klawtrade/klawtrade"
Repository = "https://github.com/klawtrade/klawtrade"
Issues = "https://github.com/klawtrade/klawtrade/issues"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "config"]
